requirements.txt
pandas
numpy  
scikit-learn
streamlit
plotly
seaborn
matplotlib
joblib
pyyaml
pytest
imbalanced-learn
